# CatalogueDemo/api-docs.json
{"swagger":"2.0","info":{"description":"DESCIPRION","version":"1.0.0","title":"Catalogue Service","termsOfService":"TERMS OF SERVICE URL","contact":{"name":"NAME","url":"URL","email":"EMAIL"},"license":{"name":"LICENSE","url":"LICENSE URL"}},"host":"localhost:8000","basePath":"/","tags":[{"name":"category-controller","description":"Category Controller"},{"name":"product-controller","description":"Product Controller"}],"paths":{"/categories/":{"get":{"tags":["category-controller"],"summary":"Get all Categories.","operationId":"retrieveAllCategoriesUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Category"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/categories/category":{"post":{"tags":["category-controller"],"summary":"Create a new Category.","operationId":"createNewCategoryUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"category","description":"category","required":true,"schema":{"$ref":"#/definitions/Category"}}],"responses":{"200":{"description":"OK","schema":{"type":"integer","format":"int64"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/categories/category/{categoryId}":{"get":{"tags":["category-controller"],"summary":"Get a Category details. 404 if the category ID is not found.","operationId":"retrieveCategoryByIdUsingGET","produces":["*/*"],"parameters":[{"name":"categoryId","in":"path","description":"categoryId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Optional«Category»"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"put":{"tags":["category-controller"],"summary":"Update a category. 404 if the category ID is not found.","operationId":"updateCategoryUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"category","description":"category","required":true,"schema":{"$ref":"#/definitions/Category"}},{"name":"categoryId","in":"path","description":"categoryId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK"},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"delete":{"tags":["category-controller"],"summary":"Delete a category. 404 if the category ID is not found.","operationId":"deleteCategoryUsingDELETE","produces":["*/*"],"parameters":[{"name":"categoryId","in":"path","description":"categoryId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"deprecated":false}},"/p1/products/CreateProduct":{"post":{"tags":["product-controller"],"summary":"create product","operationId":"CreateProductUsingPOST","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"product","description":"product","required":true,"schema":{"$ref":"#/definitions/Product"}}],"responses":{"200":{"description":"OK","schema":{"type":"string"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/p1/products/getAll":{"get":{"tags":["product-controller"],"summary":"Get all products","operationId":"retriveAllProductsUsingGET","produces":["*/*"],"responses":{"200":{"description":"OK","schema":{"type":"array","items":{"$ref":"#/definitions/Product"}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false}},"/p1/products/{productId}":{"get":{"tags":["product-controller"],"summary":"Get a product by ID","operationId":"retriveProductByIdUsingGET","produces":["*/*"],"parameters":[{"name":"productId","in":"path","description":"productId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/Optional«Product»"}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"put":{"tags":["product-controller"],"summary":"Update a product. 404 if the product is not found.","operationId":"updatePersonUsingPUT","consumes":["application/json"],"produces":["*/*"],"parameters":[{"in":"body","name":"product","description":"product","required":true,"schema":{"$ref":"#/definitions/Product"}},{"name":"productId","in":"path","description":"productId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK"},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"deprecated":false},"delete":{"tags":["product-controller"],"summary":"Delete a product. 404 if the product is not found.","operationId":"deletePersonUsingDELETE","produces":["*/*"],"parameters":[{"name":"productId","in":"path","description":"productId","required":true,"type":"integer","format":"int64"}],"responses":{"200":{"description":"OK"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}},"deprecated":false}}},"definitions":{"Category":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"}},"title":"Category"},"MonetaryAmount":{"type":"object","properties":{"currency":{"type":"string"},"value":{"type":"number"}},"title":"MonetaryAmount"},"Optional«Category»":{"type":"object","properties":{"empty":{"type":"boolean"},"present":{"type":"boolean"}},"title":"Optional«Category»"},"Optional«Product»":{"type":"object","properties":{"empty":{"type":"boolean"},"present":{"type":"boolean"}},"title":"Optional«Product»"},"Product":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string","minLength":2,"maxLength":2147483647},"price":{"$ref":"#/definitions/MonetaryAmount"},"quantity":{"type":"integer","format":"int32"},"version":{"type":"integer","format":"int32"}},"title":"Product"}}}